backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-telegram-bot==21.9
asyncpg==0.29.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2